0x03-user_authentication_service/requirements.txt
Flask==2.2.3
requests==2.25.1
SQLAlchemy==2.0.6
python_bcrypt==0.3.2